Update logic to set user instance.
comparison of the 'request.user' class with 'django.contrib.auth.models.User' class before saving the data to the Tracker model is updated with comparison from the output of get_user_model method of "django.contrib.auth"

Comparing the "request.user" with get_user_model method's output will make sure that this module works with the Django Custom User Models.

In the Tracker model for referencing the user, we are already using settings.AUTH_USER_MODEL which works well with the custom user models. This manager was the only place which was preventing the custom user model used to be counted as non-anonymous users.